Moving gzipstream module into smart_open package
There's no point keeping it in its own package anymore.
The unit tests it came with have been under smart_open/tests/ since the
beginning, anyway.
