--- gzipstream/__init__.py
@@ -1,1 +1,0 @@
-from .gzipstreamfile import GzipStreamFile
