+++ smart_open/gzipstreamfile.py
